refactor(mcp): inline start_instance + delete obsolete ops.rs module (Task #12)
Relocate start_instance from src/ops.rs into the handle_tool match arm in
src/mcp/handlers.rs and delete the ops module entirely. start_instance
was the last remaining function in ops.rs after Task #9 Option C/PR #18
removed the 21 orphaned CLI-wrapper functions (CLI migrated to direct
api::call months ago, leaving ops wrappers dead code masked by
#![allow(dead_code)]).

Form chosen: inline match-arm body — matches the established convention
of all 4 peer instance operations (delete_instance, describe_instance,
replace_instance, set_display_name), which are each inlined directly
into handle_tool match arms. No private helper introduced because no
peer uses that form for single-caller instance ops.

YAGNI: rejected creating an agent_lifecycle.rs sibling module. With only
one function, extraction creates indirection without cohesion — colocating
with peer instance handlers in handle_tool is the simplest structure that
matches current reality.

Dropped both ops.rs unit tests:
- no_mcp_prefix_in_ops: static file-content assert on a deleted file
- backend_resolves_to_preset_command: near-duplicate of
  backend::from_command_detection (src/backend.rs:463)

Test count: 621 → 619 (two dead tests removed, no behavior tests lost).
CI gates: cargo fmt clean, cargo clippy --all-targets -D warnings clean,
cargo test 619 passed / 0 failed.
