+++ src/mcp/handlers.rs
@@ -379,7 +379,43 @@ pub fn handle_tool(tool: &str, args: &Value, instance_name: &str) -> Value {
             }
             json!({"name": name})
         }
-        "start_instance" => crate::ops::start_instance(&home, args),
+        "start_instance" => {
+            let name = match args["name"].as_str() {
+                Some(n) => n,
+                None => return json!({"error": "missing 'name'"}),
+            };
+            if let Err(e) = crate::agent::validate_name(name) {
+                return json!({"error": e});
+            }
+            let fleet_path = home.join("fleet.yaml");
+            if !fleet_path.exists() {
+                return json!({"error": "No fleet.yaml"});
+            }
+            let config = match crate::fleet::FleetConfig::load(&fleet_path) {
+                Ok(c) => c,
+                Err(e) => return json!({"error": format!("fleet.yaml: {e}")}),
+            };
+            match config.resolve_instance(name) {
+                Some(resolved) => {
+                    let cmd_args = resolved.args.join(" ");
+                    match crate::api::call(
+                        &home,
+                        &json!({"method": crate::api::method::SPAWN, "params": {
+                            "name": name, "backend": resolved.backend_command, "args": cmd_args,
+                            "mode": "resume",
+                            "working_directory": resolved.working_directory.map(|p| p.display().to_string()),
+                        }}),
+                    ) {
+                        Ok(resp) if resp["ok"].as_bool() == Some(true) => json!({"name": name}),
+                        Ok(resp) => {
+                            json!({"error": resp["error"].as_str().unwrap_or("spawn failed")})
+                        }
+                        Err(e) => json!({"error": format!("API unavailable: {e}")}),
+                    }
+                }
+                None => json!({"error": format!("Instance '{name}' not in fleet.yaml")}),
+            }
+        }
         "describe_instance" => {
             let name = args["name"].as_str().unwrap_or("");
             if let Err(e) = crate::agent::validate_name(name) {
